Split out the discovery related routes, and fix symbol visibility in pdns
