+++ src/routes.rs
@@ -4,6 +4,7 @@
 
 use config::Config;
 use database::{DatabaseError, DomainRecord};
+use discovery::{adddiscovery, discovery, ping, revokediscovery};
 use errors::*;
 use iron::headers::ContentType;
 use iron::prelude::*;
@@ -17,12 +18,6 @@ use uuid::Uuid;
 
 fn domain_for_name(name: &str, config: &Config) -> String {
     format!("{}.box.{}.", name, config.domain).to_lowercase()
-}
-
-#[derive(Serialize)]
-struct Discovered {
-    href: String,
-    desc: String,
 }
 
 fn register(req: &mut Request, config: &Config) -> IronResult<Response> {
@@ -103,39 +98,6 @@ fn info(req: &mut Request, config: &Config) -> IronResult<Response> {
     match config.db.get_record_by_token(&token).recv().unwrap() {
         Ok(record) => {
             let mut response = Response::with(serde_json::to_string(&record).unwrap());
-            response.headers.set(ContentType::json());
-            Ok(response)
-        }
-        Err(DatabaseError::NoRecord) => EndpointError::with(status::BadRequest, 400),
-        Err(_) => EndpointError::with(status::InternalServerError, 501),
-    }
-}
-
-// Public ping endpoint, returning names of servers on the same
-// local network than the client.
-fn ping(req: &mut Request, config: &Config) -> IronResult<Response> {
-    info!("GET /ping");
-
-    let remote_ip = format!("{}", req.remote_addr.ip());
-
-    match config
-              .db
-              .get_records_by_public_ip(&remote_ip)
-              .recv()
-              .unwrap() {
-        Ok(records) => {
-            let results: Vec<Discovered> = records
-                .into_iter()
-                .map(|item| {
-                         Discovered {
-                             href: format!("https://{}",
-                                           item.local_name[..item.local_name.len() - 1].to_owned()),
-                             desc: item.description,
-                         }
-                     })
-                .collect();
-
-            let mut response = Response::with(serde_json::to_string(&results).unwrap());
             response.headers.set(ContentType::json());
             Ok(response)
         }
@@ -298,140 +260,6 @@ fn dnsconfig(req: &mut Request, config: &Config) -> IronResult<Response> {
                     Ok(response)
                 }
                 Err(_) => EndpointError::with(status::InternalServerError, 501),
-            }
-        }
-        Err(DatabaseError::NoRecord) => EndpointError::with(status::BadRequest, 400),
-        Err(_) => EndpointError::with(status::InternalServerError, 501),
-    }
-}
-
-fn adddiscovery(req: &mut Request, config: &Config) -> IronResult<Response> {
-    info!("GET /adddiscovery");
-
-    let map = req.get_ref::<Params>().unwrap(); // TODO: don't unwrap.
-    let token = map.find(&["token"]);
-    let disco = map.find(&["disco"]);
-
-    if token.is_none() || disco.is_none() {
-        return EndpointError::with(status::BadRequest, 400);
-    }
-
-    let token = String::from_value(token.unwrap()).unwrap();
-    let disco = String::from_value(disco.unwrap()).unwrap();
-
-    match config.db.get_record_by_token(&token).recv().unwrap() {
-        Ok(_) => {
-            match config.db.add_discovery(&token, &disco).recv().unwrap() {
-                Ok(()) => {
-                    let mut response = Response::new();
-                    response.status = Some(Status::Ok);
-
-                    Ok(response)
-                }
-                Err(_) => EndpointError::with(status::BadRequest, 400),
-            }
-        }
-        Err(DatabaseError::NoRecord) => EndpointError::with(status::BadRequest, 400),
-        Err(_) => EndpointError::with(status::InternalServerError, 501),
-    }
-}
-
-fn revokediscovery(req: &mut Request, config: &Config) -> IronResult<Response> {
-    info!("GET /revokediscovery");
-
-    let map = req.get_ref::<Params>().unwrap(); // TODO: don't unwrap.
-    let token = map.find(&["token"]);
-    let disco = map.find(&["disco"]);
-
-    if token.is_none() || disco.is_none() {
-        return EndpointError::with(status::BadRequest, 400);
-    }
-
-    let token = String::from_value(token.unwrap()).unwrap();
-    let disco = String::from_value(disco.unwrap()).unwrap();
-
-    match config.db.get_record_by_token(&token).recv().unwrap() {
-        Ok(_) => {
-            match config.db.delete_discovery(&disco).recv().unwrap() {
-                Ok(_) => {
-                    let mut response = Response::new();
-                    response.status = Some(Status::Ok);
-
-                    Ok(response)
-                }
-                Err(_) => EndpointError::with(status::BadRequest, 400),
-            }
-        }
-        Err(DatabaseError::NoRecord) => EndpointError::with(status::BadRequest, 400),
-        Err(_) => EndpointError::with(status::InternalServerError, 501),
-    }
-}
-
-fn discovery(req: &mut Request, config: &Config) -> IronResult<Response> {
-    info!("GET /discovery");
-
-    let remote_ip = format!("{}", req.remote_addr.ip());
-
-    let map = req.get_ref::<Params>().unwrap(); // TODO: don't unwrap.
-    let disco = map.find(&["disco"]);
-
-    if disco.is_none() {
-        return EndpointError::with(status::BadRequest, 400);
-    }
-
-    let disco = String::from_value(disco.unwrap()).unwrap();
-
-    match config
-              .db
-              .get_token_for_discovery(&disco)
-              .recv()
-              .unwrap() {
-        Ok(token) => {
-            match config
-                      .db
-                      .get_records_by_public_ip(&remote_ip)
-                      .recv()
-                      .unwrap() {
-                Ok(records) => {
-                    // Filter out and only return the record that matches the token.
-                    let results: Vec<Discovered> = records
-                        .into_iter()
-                        .filter(|item| item.token == token)
-                        .map(|item| {
-                                 Discovered {
-                                     href: format!("https://{}",
-                                                   item.local_name[..item.local_name.len() - 1]
-                                                       .to_owned()),
-                                     desc: item.description,
-                                 }
-                             })
-                        .collect();
-
-                    if results.is_empty() {
-                        // If the result vector is empty, return the remote name for this token.
-                        match config.db.get_record_by_token(&token).recv().unwrap() {
-                            Ok(record) => {
-                                let len = record.remote_name.len() - 1;
-                                let result = vec![Discovered {
-                                                      href: format!("https://{}",
-                                                                    record.remote_name[..len]
-                                                                        .to_owned()),
-                                                      desc: record.description,
-                                                  }];
-                                let mut response = Response::with(serde_json::to_string(&result)
-                                                                      .unwrap());
-                                response.headers.set(ContentType::json());
-                                Ok(response)
-                            }
-                            Err(_) => EndpointError::with(status::BadRequest, 400),
-                        }
-                    } else {
-                        let mut response = Response::with(serde_json::to_string(&results).unwrap());
-                        response.headers.set(ContentType::json());
-                        Ok(response)
-                    }
-                }
-                Err(_) => EndpointError::with(status::BadRequest, 400),
             }
         }
         Err(DatabaseError::NoRecord) => EndpointError::with(status::BadRequest, 400),
